Intern_Submissions/Palla_Umadevi/requirements.txt
django>=5.0
djangorestframework

torch
openai-whisper

librosa
numpy
scipy
soundfile
pydub
noisereduce
ffmpeg-python

sentence-transformers
transformers
keybert
scikit-learn
nltk

jiwer
contractions
